Split classes out into more manageable files and add demo machine
